+++ src/Components/Input/index.js
@@ -1,5 +1,6 @@
 import BooleanInput from './BooleanInput';
 import DateTimeInput from './DateTimeInput';
+import DepartmentInput from './DepartmentInput';
 import DurationInput from './DurationInput';
 import NumericRangeInput from './NumericRangeInput';
 import SelectInput from './SelectInput';
@@ -8,6 +9,7 @@ import TextInput from './TextInput';
 export {
   BooleanInput,
   DateTimeInput,
+  DepartmentInput,
   DurationInput,
   NumericRangeInput,
   SelectInput,