+++ src/Components/index.js
@@ -1,4 +1,6 @@
+import BooleanInput from './Input/BooleanInput';
 import DateTimeInput from './Input/DateTimeInput';
+import DepartmentInput from './Input/DepartmentInput';
 import DurationInput from './Input/DurationInput';
 import NumericRangeInput from './Input/NumericRangeInput';
 import SelectInput from './Input/SelectInput';
@@ -6,7 +8,9 @@ import TextInput from './Input/TextInput';
 import TokenField from './TokenField';
 
 export {
+  BooleanInput,
   DateTimeInput,
+  DepartmentInput,
   DurationInput,
   NumericRangeInput,
   SelectInput,