+++ tests/storybook/components/TokenField.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Immutable from 'immutable';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import TokenField from 'Components/TokenField';
@@ -36,6 +37,121 @@ const countries = [
   { label: 'Sweden', value: 'SE' },
   { label: 'United Kingdom', value: 'GB' }
 ];
+
+const departments = {
+  1: {
+    user_title:         'Support',
+    parent:             null,
+    brands:             [1],
+    is_chat_enabled:    false,
+    display_order:      0,
+    is_tickets_enabled: true,
+    title:              'Support',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       1,
+    children: []
+  },
+  2: {
+    user_title:         'Sales',
+    parent:             null,
+    brands:             [1],
+    is_chat_enabled:    false,
+    display_order:      0,
+    is_tickets_enabled: true,
+    title:              'Sales',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       2,
+    children: []
+  },
+  5: {
+    user_title:         'Widgets',
+    parent:             null,
+    brands:             [1],
+    is_chat_enabled:    false,
+    display_order:      0,
+    is_tickets_enabled: true,
+    title:              'Widgets',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       5,
+    children: []
+  },
+  6: {
+    user_title:         'Regulation and Control of Magical Creatures',
+    parent:             null,
+    brands:             [],
+    is_chat_enabled:    false,
+    display_order:      1,
+    is_tickets_enabled: true,
+    title:              'Regulation and Control of Magical Creatures',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       6,
+    children: [7, 8]
+  },
+  7: {
+    user_title:         'Regulation',
+    parent:             6,
+    brands:             [1],
+    is_chat_enabled:    false,
+    display_order:      2,
+    is_tickets_enabled: true,
+    title:              'Regulation',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       7,
+    children: []
+  },
+  8: {
+    user_title:         'Control',
+    parent:             6,
+    brands:             [1],
+    is_chat_enabled:    false,
+    display_order:      3,
+    is_tickets_enabled: true,
+    title:              'Control',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       8,
+    children: []
+  },
+  9: {
+    user_title:         'Hotdogs',
+    parent:             null,
+    brands:             [1],
+    is_chat_enabled:    false,
+    display_order:      4,
+    is_tickets_enabled: true,
+    title:              'Hotdogs',
+    avatar:             {
+      default_url_pattern: 'http://deskpro5.local/file.php/o-avatar/default?s={{IMG_SIZE}}&size-fit=1',
+      url_pattern:         null,
+      base_gravatar_url:   null
+    },
+    id:       9,
+    children: []
+  }
+};
 
 const fakeAPI = (filter, delay, value) => new Promise(((resolve) => {
   setTimeout(resolve, delay, value.filter(o => o.title === filter || filter === ''));
@@ -119,6 +235,15 @@ const tokenTypes = [
       renderHeader: <h3>Countries</h3>,
       showSearch:   false
     },
+  },
+  {
+    id:     'department',
+    widget: 'DepartmentInput',
+    props:  {
+      dataSource: {
+        getOptions: Immutable.fromJS(departments)
+      }
+    }
   },
   {
     id:     'country',
@@ -208,7 +333,6 @@ class TokenFieldStory extends React.Component {
           onBlur={action('Blur')}
           placeholder="Search ..."
           showTokensOnFocus={this.props.showTokensOnFocus}
-          popupOpen
         />
         <br />
         <Button onClick={this.addCountryToken}>Add country token</Button>