+++ tests/storybook/index.js
@@ -2,6 +2,7 @@ import '@deskpro/react-components/dist/main.css';
 import './components/TokenField';
 import './components/input/BooleanInput';
 import './components/input/DateTimeInput';
+import './components/input/DepartmentInput';
 import './components/input/DurationInput';
 import './components/input/NumericRangeInput';
 import './components/input/SelectInput';